chore: replace editor
